// app/src/types/modules/financial-maintenance/discount-management/DiscountManagementTypes.ts
export type DiscountType = "Percentage" | "Fixed";
export type DiscountTransactionType = "Purchase" | "Sales";
export type DiscountStatus = "Active" | "Inactive";
export type DiscountTypeFilter = "All" | DiscountTransactionType;
export type DiscountValueTypeFilter = "All" | DiscountType;
export type DiscountStatusFilter = "" | DiscountStatus;

export type Discount = {
	id: string;
	name: string;
	description: string;
	type: DiscountTransactionType;
	discountType: DiscountType;
	amount: number;
	status: DiscountStatus;
	accountId?: string;
	accountCode?: string;
	accountTitle?: string;
	accountGroupPath?: string;
	createdBy?: string | null;
	createdAt?: string;
	updatedBy?: string | null;
	updatedAt?: string;
};